cc_util rename
