+++ SMAC/util/maa2c_tools.py
@@ -1,9 +1,10 @@
+
 from ray.rllib.agents.a3c.a3c_tf_policy import actor_critic_loss as tf_loss
 from ray.rllib.agents.a3c.a3c_torch_policy import actor_critic_loss as torch_loss
 from ray.rllib.evaluation.postprocessing import Postprocessing
 from ray.rllib.utils.tf_ops import explained_variance
 from ray.rllib.utils.framework import try_import_tf, try_import_torch
-from MPE.cc_utils.mappo_tools import CentralizedValueMixin
+from SMAC.util.mappo_tools import CentralizedValueMixin
 
 
 tf1, tf, tfv = try_import_tf()
@@ -17,12 +18,12 @@ def loss_with_central_critic_a2c(policy, model, dist_class, train_batch):
 
     vf_saved = model.value_function
     model.value_function = lambda: policy.model.central_value_function(
-        train_batch["obs"], train_batch["opponent_obs"],
-        train_batch["opponent_action"])
+        train_batch["self_obs"], train_batch["state"])
 
+    # recording data
     policy._central_value_out = model.value_function()
-    loss = func(policy, model, dist_class, train_batch)
 
+    loss = func(policy, model, dist_class, train_batch)
     model.value_function = vf_saved
 
     return loss
